fixes problem with multiple aud
In some cases, auth0 includes a second aud for userinfo endpoint.  Tokens can have multiple target audiences as long as the custom API’s signing algorithm is set to RS256.  Works with auth0 if machine-to-machine api userinfo is added as a second aud.  This patch prevents JWT validation from failing in that case.  See https://community.auth0.com/t/my-token-has-multiple-audiences-is-that-normal/41417